Move data cleaning fns to utilities file
